Module 35.3

+++ src/order/order.controller.ts
@@ -1,41 +1,52 @@
 /* eslint-disable prettier/prettier */
-import { Body, Controller, Delete, Get, NotFoundException, Param, Post, Put } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Param,
+  Delete,
+  Body,
+  Post,
+  Put,
+} from '@nestjs/common';
 import { OrdersService } from './order.service';
-import { CreateOrderDTO } from './../products/dtos/create-order.dto';
-import { UpdateOrderDTO } from './../products/dtos/update-order.dto';
+import { NotFoundException } from '@nestjs/common';
 import { ParseUUIDPipe } from '@nestjs/common';
+import { CreateOrderDTO } from 'src/products/dtos/create-order.dto';
+import { UpdateOrderDTO } from 'src/products/dtos/update-order.dto';
 
 @Controller('orders')
 export class OrdersController {
   constructor(private ordersService: OrdersService) {}
-
   @Get('/')
-  async getAll() {
+  getAll(): any {
     return this.ordersService.getAll();
   }
-
   @Get('/:id')
   async getById(@Param('id', new ParseUUIDPipe()) id: string) {
-    const order = await this.ordersService.getById(id);
-    if (!order) throw new NotFoundException('Order not found');
-    return order;
+    const ord = await this.ordersService.getById(id);
+    if (!ord) throw new NotFoundException('Order not found');
+    return ord;
   }
-
+  @Delete('/:id')
+  async deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!(await this.ordersService.getById(id)))
+      throw new NotFoundException('Order not found');
+    await this.ordersService.deleteById(id);
+    return { success: true };
+  }
   @Post('/')
-  async create(@Body() orderData: CreateOrderDTO) {
+  create(@Body() orderData: CreateOrderDTO) {
     return this.ordersService.create(orderData);
   }
-
   @Put('/:id')
   async update(
     @Param('id', new ParseUUIDPipe()) id: string,
     @Body() orderData: UpdateOrderDTO,
   ) {
-    return this.ordersService.updateById(id, orderData);
-  }
+    if (!(await this.ordersService.getById(id)))
+      throw new NotFoundException('Order not found');
 
-  @Delete('/:id')
-  async deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
-    return this.ordersService.deleteById(id);
+    await this.ordersService.updateById(id, orderData);
+    return { success: true };
   }
 }
